+++ lib/scripts/banner.py
@@ -1,5 +1,5 @@
 
-__version__ = '1.0.7'
+__version__ = '1.0.8'
 
 
 def show_banner():