+++ changelog.md
@@ -14,10 +14,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     - Added `get_azurecreds` command to pull Azure co-management application blobs
     - Added `get_azuretenant` commant to pull Azure tenant info
     - Added `get_pxepassword` command to pull PXE boot blobs if configured
-    - Added `decrypt` command to decrypt passed credential blob 
+    - Added `get_forestkey` command to pull forest discovery session key blobs
+    - Added `decrypt` command to decrypt passed credential blob
+    - Added `decryptEx` command to decrypt forest discovery credential blobs
         - You've got to be "interactive" with the SCCM primary site server for decryption to work 
         - This means the site server must be a client 
         - Uses script execution
+    - Updates thanks to [Parzel](https://bsky.app/profile/parzel.bsky.social):
+    - Added `list_script` command to list scripts published to SCCM
+    - Added `delete_script` commmand to delete a target script from SCCM 
    
 ## [1.0.7] - 2025-02-28